// back/src/common/enums/external-APIs/mail.provider.enum.ts
export enum MailTemplateIds {
  // Auth
  USER_VALIDATION = 'd-efe80bc5899849aab747af557fd34fa0',
  PASSWORD_RECOVERY = 'd-27739544fa5f4034804cd904ba203db4',
  MAIL_CHANGE_REQUEST_NEW = 'd-8921a090948e4af6bfd7fccc0f3206ff',
  MAIL_CHANGE_REQUEST_OLD = 'd-e48f0ce11bfd4571ba15957463a4db56',

  // Bookings
  BOOKING_CANCELED = 'd-69f6431662894e81b4375d5270b8e93d',
  BOOKING_VALIDATED = 'd-aa837fec34cd47748e328180bdaaa15c',
  BOOKING_CREATED_TO_USER = 'd-b32aaec5608940bcb2ff6e7eb4ea82ec',
  BOOKING_CREATED_TO_ADMIN = 'd-478df834c6b4463a841f2cc85f370001',
  // Missions
  MISSION_CREATED_TO_CLIENT = 'd-8dec2a01c6f9492bb5a52f767c6c10bf',
  MISSION_RECEIVED_TO_PROFESSIONAL = 'd-079eb0638cc24318a34571e74b77a256',
  MISSION_CANCELED_TO_CLIENT = 'd-1a3bc6be963d41b3a536c01b475e752b',
  MISSION_CANCELED_TO_PROFESSIONAL = 'd-6a30af66023c47cf82009e8015bb16a9',
  MISSION_PROFESSIONAL_WITH_PENDING_QUOTE_REFUSED_MISSION_TO_CLIENT = 'd-5cd9648161044ea4b1f845969daa723c',
  MISSION_NEW_PENDING_QUOTE_RECEIVED_TO_CLIENT = 'd-5ec0b351c0e64b8585c306b2f19faa53',
  MISSION_NEW_MESSAGE = 'd-b473c13867d3410e8f97f70e2517c4a7',
  MISSION_CLIENT_CHOSE_ANOTHER_PROFESSIONAL = 'd-a019afa4361c4e2186290c684f0fc949',
  MISSION_COMPLETED_TO_CLIENT = 'd-9c6ad62af9ea4e25b5dbc6649a07f853',
  MISSION_VALIDATED_BY_CLIENT_TO_CLIENT = 'd-c642ffdeb6784190a7485f28dd3cd0bc',
  MISSION_VALIDATED_BY_CLIENT_TO_PROFESSIONAL = 'd-f6b931d1301b40c4a9b012d94114fa7b',
  MISSION_CLIENT_OPENED_DISPUTE_TO_ADMIN = 'd-692065195526429bb4188096b33ecd42',
}